+++ frontend/views/oauth/create_oauth.php
@@ -7,7 +7,8 @@
     use frontend\models\AppSocialLogin;
 
     $this->title = Yii::$app->name . ' | ' . Yii::t('common', 'OAuth2 configuration');
-    $this->params['breadcrumbs'][] = ['label' => Yii::t('common', 'Developer'), 'url' => ['/developer/index']];
+    $this->params['breadcrumbs'][] = ['label' => Yii::t('common', 'Developer')];
+    $this->params['breadcrumbs'][] = ['label' => Yii::t('common', 'My apps'), 'url' => ['developer/index']];
     $this->params['breadcrumbs'][] = ['label' => $app->name, 'url' => ['/developer/details', 'id' => $app->id]];
     $this->params['breadcrumbs'][] = Yii::t('common', 'OAuth2 configuration');